uefi: mem: decouple memory_map/mod.rs into multiple smaller modules
